refactor: internal-app -> core repository 이동
