--- matzip-app-internal-api/src/main/java/com/woowacourse/matzip/repository/CampusRepository.java
@@ -1,7 +1,0 @@
-package com.woowacourse.matzip.repository;
-
-import com.woowacourse.matzip.domain.campus.Campus;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CampusRepository extends JpaRepository<Campus, Long> {
-}
