--- matzip-app-internal-api/src/main/java/com/woowacourse/matzip/repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.woowacourse.matzip.repository;
-
-import com.woowacourse.matzip.domain.category.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
