--- matzip-app-internal-api/src/main/java/com/woowacourse/matzip/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package com.woowacourse.matzip.repository;
-
-import com.woowacourse.matzip.domain.member.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
