--- matzip-app-internal-api/src/main/java/com/woowacourse/matzip/repository/RestaurantRepository.java
@@ -1,7 +1,0 @@
-package com.woowacourse.matzip.repository;
-
-import com.woowacourse.matzip.domain.restaurant.Restaurant;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
-}
