--- matzip-app-internal-api/src/main/java/com/woowacourse/matzip/repository/ReviewRepository.java
@@ -1,7 +1,0 @@
-package com.woowacourse.matzip.repository;
-
-import com.woowacourse.matzip.domain.review.Review;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ReviewRepository extends JpaRepository<Review, Long> {
-}
